moved auth_headers fixture to confest
