+++ api/tests/test_pantry_entry_routes.py
@@ -1,0 +1,2 @@
+import pytest
+from app import db
